EC-513 Refactor of document processing
